Feature/add crop image

+++ Eshop/EshopSolution.Application/Cacalog/Products/CategoryService.cs
@@ -1,4 +1,5 @@
 ﻿using EshopSolution.Data.EF;
+using EshopSolution.Data.Enums;
 using EshopSolution.Data.Entities;
 using EshopSolution.ViewModels.Catalog.Categories;
 using EshopSolution.ViewModels.Common;
@@ -33,6 +34,7 @@ namespace EshopSolution.Application.Cacalog.Products
             query = query.Skip((request.PageIndex - 1) * request.PageSize);
             var data = query.Select(x => new CategoryViewModel()
             {
+                
                 Id = x.c.Id,
                 Name = x.ct.Name,
                 IsShowOnHome = x.c.IsShowOnHome,
@@ -65,6 +67,9 @@ namespace EshopSolution.Application.Cacalog.Products
             //2.filer by languageid
             query = query.Where(x => x.ct.LanguageId == request.LanguageId);
 
+            //filter by status
+            if(request.Status!=null)
+            query = query.Where(x => x.c.Status == request.Status);
             //3/filter by keyword
             if (!String.IsNullOrWhiteSpace(request.Keyword))
                 query = query.Where(x => x.ct.Name.Contains(request.Keyword)
@@ -228,14 +233,22 @@ namespace EshopSolution.Application.Cacalog.Products
         {
             Category category = _context.Categories.Where(x => x.Id == id).FirstOrDefault();
             if (category == null)
-                return new ApiErrorResult<bool>();
+                return new ApiErrorResult<bool>(true);
 
-            _context.Categories.Remove(category);
-
-            if (await _context.SaveChangesAsync() > 0)
-                return new ApiSuccessResult<bool>();
+            if (_context.ProductInCategories.Where(x => x.CategoryId == id).First() == null)
+            {
+                _context.Categories.Remove(category);
+                await _context.SaveChangesAsync();
+                return new ApiSuccessResult<bool>("Xóa thành công", true);
+            }
             else
-                return new ApiErrorResult<bool>();
+            {
+                category.Status = Status.InActive;
+                await _context.SaveChangesAsync();
+                return new ApiSuccessResult<bool>("Danh mục đã được vô hiệu hóa", true);
+            }
+              
+
         }
     }
 }
